refactor(memory): Extract shared EmbeddingModelInfo class
Remove duplicate inner EmbeddingModelInfo classes from
BedrockEmbeddingModelProvider and OpenaiEmbeddingModelProvider.
Create shared EmbeddingModelInfo in memorycontainer package.
