+++ common/src/main/java/org/opensearch/ml/common/agent/BedrockEmbeddingModelProvider.java
@@ -18,6 +18,7 @@ import org.opensearch.ml.common.connector.ConnectorClientConfig;
 import org.opensearch.ml.common.connector.ConnectorProtocols;
 import org.opensearch.ml.common.input.execute.agent.ContentBlock;
 import org.opensearch.ml.common.input.execute.agent.Message;
+import org.opensearch.ml.common.memorycontainer.EmbeddingModelInfo;
 import org.opensearch.ml.common.memorycontainer.MemoryContainerConstants;
 import org.opensearch.ml.common.model.ModelProvider;
 import org.opensearch.ml.common.transport.register.MLRegisterModelInput;
@@ -180,16 +181,4 @@ public class BedrockEmbeddingModelProvider extends ModelProvider {
         return KNOWN_MODELS.get(modelId);
     }
 
-    /**
-     * Info about a known embedding model.
-     */
-    public static class EmbeddingModelInfo {
-        public final FunctionName functionName;
-        public final int dimension;
-
-        public EmbeddingModelInfo(FunctionName functionName, int dimension) {
-            this.functionName = functionName;
-            this.dimension = dimension;
-        }
-    }
 }
